+++ optimalportfolios/examples/covar_estimation/demo_covar_different_estimation_freqs.py
@@ -35,7 +35,7 @@ ac_group_order = ['Equity', 'Bonds', 'Bonds']
 
 # set lasso model
 lasso_params = dict(reg_lambda=1e-5, span=120, demean=False, solver='CLARABEL', warmup_period=50)
-lasso_model = LassoModel(model_type=LassoModelType.GROUP_LASSO_CLUSTERS, **lasso_params)
+lasso_model = LassoModel(model_type=LassoModelType.HIERARCHICAL_CLUSTER_GROUP_LASSO, **lasso_params)
 
 
 estimator_daily = FactorCovarEstimator(lasso_model=LassoModel(),